requirements.txt
Authlib==1.3.2
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
flask-talisman==1.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jose==1.0.0
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
